% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.r
\name{col_labels}
\alias{col_labels}
\title{Calculate column labels.}
\usage{
col_labels(df)
}
\arguments{
\item{df}{data frame produced by \code{\link{prodcalc}}}
}
\description{
Calculate column labels.
}
\keyword{internal}
