% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.r
\name{find_col_level}
\alias{find_col_level}
\title{Find the first level which has columns.}
\usage{
find_col_level(df)
}
\arguments{
\item{df}{data frame of rectangle positions}
}
\description{
Returns \code{NA} if no columns at any level.
}

