% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.r
\name{flucts}
\alias{flucts}
\title{Template for a fluctuation diagram.}
\usage{
flucts(direction = "h")
}
\arguments{
\item{direction}{direction of first split}
}
\description{
Template for a fluctuation diagram.
}
