\name{hbar}
\alias{hbar}
\title{Horizontal bar partition: width constant, height varies.}
\usage{hbar(data, bounds, offset=0.02, max)}

\description{
  Horizontal bar partition: width constant, height varies.
}
\arguments{
  \item{data}{bounds data frame}
  \item{bounds}{bounds of space to partition}
  \item{offset}{space between spines}
  \item{max}{maximum value}
}
