% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.r
\name{nested}
\alias{nested}
\title{Template for a nested barchart.
A nested bar is just a sequence of bars in the same direction.}
\usage{
nested(direction = "h")
}
\arguments{
\item{direction}{direction of first split}
}
\description{
Template for a nested barchart.
A nested bar is just a sequence of bars in the same direction.
}

