% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.r
\name{row_labels}
\alias{row_labels}
\title{Calculate row labels.}
\usage{
row_labels(df)
}
\arguments{
\item{df}{data frame produced by \code{\link{prodcalc}}}
}
\description{
Calculate row labels.
}
\keyword{internal}
