% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.r
\name{scale_x_product}
\alias{scale_x_product}
\title{Generate an x-scale for ggplot2 graphics.}
\usage{
scale_x_product(df)
}
\arguments{
\item{df}{list of data frame produced by \code{\link{prodcalc}}, formula and divider}
}
\description{
Generate an x-scale for ggplot2 graphics.
}

