\name{spine}
\alias{spine}
\title{Spine partition: divide longest dimesion.}
\usage{
  spine(data, bounds, offset = 0.01, max = NULL)
}
\arguments{
  \item{data}{bounds data frame}

  \item{bounds}{bounds of space to partition}

  \item{offset}{space between spines}

  \item{max}{maximum value}
}
\description{
  Spine partition: divide longest dimesion.
}

