% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.r
\name{stacked}
\alias{stacked}
\title{Template for a stacked bar chart.
A stacked bar chart starts with a bar and then continues with spines in the
opposite direction.}
\usage{
stacked(direction = "h")
}
\arguments{
\item{direction}{direction of first split}
}
\description{
Template for a stacked bar chart.
A stacked bar chart starts with a bar and then continues with spines in the
opposite direction.
}
