\name{vbar}
\alias{vbar}
\title{Vertical bar partition: height constant, width varies.}
\usage{
  vbar(data, bounds, offset = 0.02, max = NULL)
}
\arguments{
  \item{data}{bounds data frame}

  \item{bounds}{bounds of space to partition}

  \item{offset}{space between spines}

  \item{max}{maximum value}
}
\description{
  Vertical bar partition: height constant, width varies.
}

