\name{find_col_level}
\alias{find_col_level}
\title{Find the first level which has columns.}
\usage{find_col_level(df)}

\description{
  Find the first level which has columns.
}

\details{
  Returns \code{NA} if no columns at any level.
}
\arguments{
  \item{df}{data frame of rectangle positions}
}
