% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition-2d.r
\name{fluct}
\alias{fluct}
\title{Fluctation partitioning.}
\usage{
fluct(data, bounds, offset = 0.05, max = NULL)
}
\arguments{
\item{data}{bounds data frame}

\item{bounds}{bounds of space to partition}

\item{offset}{space between spines}

\item{max}{maximum value}
}
\description{
Fluctation partitioning.
}
