\name{fluct}
\alias{fluct}
\title{Fluctation partitioning.}
\usage{fluct(data, bounds, offset=0.05, max)}

\description{
  Fluctation partitioning.
}
\arguments{
  \item{data}{bounds data frame}
  \item{bounds}{bounds of space to partition}
  \item{offset}{space between spines}
  \item{max}{maximum value}
}
