\name{parse_product_formula}
\alias{parse_product_formula}
\title{Parse product formula into component pieces}
\usage{
  parse_product_formula(f)
}
\arguments{
  \item{f}{function to parse into component pieces}
}
\value{
  \item{wt}{the weighting variable} \item{cond}{condition
  variables} \item{margin}{margining variables}
}
\description{
  Parse product formula into component pieces
}
\keyword{internal}

