\name{tile}
\alias{tile}
\title{Tree map partitioning.}
\usage{
  tile(data, bounds, max = 1)
}
\arguments{
  \item{data}{bounds data frame}

  \item{bounds}{bounds of space to partition}

  \item{max}{maximum value}
}
\description{
  Adapated from SquarifiedLayout in
  \url{http://www.cs.umd.edu/hcil/treemap-history/Treemaps-Java-Algorithms.zip}
}

